requirements.txt
discord.py>=2.3,<3.0
google-genai>=1.0,<2.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
tavily-python>=0.5,<1.0
numpy>=1.24,<3.0
